backend/requirements.txt
# Let google-adk pull compatible fastapi, uvicorn, websockets
google-adk>=1.20.0
python-dotenv==1.0.0
pydantic-settings>=2.5.2
Pillow==10.2.0
httpx>=0.27.0
